Split backend_auth.rs into config/startup/live + expand tests/support
Extracts shared helpers into tests/support/{fixtures,http_mock,waits}.rs
and distributes the 6 tests across backend_auth_config (3 sync env-var
tests), backend_auth_startup (2 async mock-HTTP tests), and
backend_auth_live (1 ignored live-API test).
